+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/_params.py
@@ -79,6 +79,7 @@ def load_arguments(self, _):
 
     with self.argument_context('identity create') as c:
         c.argument('location', get_location_type(self.cli_ctx))
+        c.argument('tags', tags_type)
     # endregion
 
     # region Snapshots