+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/_validators.py
@@ -1254,4 +1254,5 @@ def process_remove_identity_namespace(cmd, namespace):
 # TODO move to its own command module https://github.com/Azure/azure-cli/issues/5105
 def process_msi_namespace(cmd, namespace):
     get_default_location_from_resource_group(cmd, namespace)
+    validate_tags(namespace)
 # endregion